Add files via upload
This file contains a simple python code for creating a random walk. This code is very useful if you need to create a synthetic dataset for your proposed framework or technique to test. It can be effectively used to create a series of data and then you can decide which to keep by setting simple threshold value.